refactor: move getPollContent to its own file to avoid circular dep
